+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/mealredesign/ui/MealsCardsViewModel.kt
@@ -1,25 +1,37 @@
 package com.maksimowiczm.foodyou.feature.mealredesign.ui
 
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.maksimowiczm.foodyou.core.domain.repository.MealRepository
 import com.maksimowiczm.foodyou.core.domain.repository.MeasurementRepository
 import com.maksimowiczm.foodyou.core.ext.combine
 import com.maksimowiczm.foodyou.core.ext.launch
+import com.maksimowiczm.foodyou.feature.meal.data.observeMealCardsLayout
 import com.maksimowiczm.foodyou.feature.mealredesign.domain.Meal
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.runBlocking
 import kotlinx.datetime.LocalDate
 
 internal class MealsCardsViewModel(
     private val mealRepository: MealRepository,
-    private val measurementRepository: MeasurementRepository
+    private val measurementRepository: MeasurementRepository,
+    private val dataStore: DataStore<Preferences>
 ) : ViewModel() {
+
+    val layout = dataStore.observeMealCardsLayout().stateIn(
+        scope = viewModelScope,
+        started = SharingStarted.WhileSubscribed(2_000),
+        initialValue = runBlocking { dataStore.observeMealCardsLayout().first() }
+    )
 
     private val dateState = MutableStateFlow<LocalDate?>(null)
